ci: limit running the kernel tests to the UML environment
The kernel tests have additional kernel and userspace requirements.  For
example,

- Loading IMA policy rules requires the ability of writing to the IMA
policy file <securityfs>/security/integrity/ima/policy.

- Loading additional IMA policy rules might require the rules to be
signed.  This requires access to the private key associated with the IMA
code signing key loaded on the IMA keyring.

- The fsverity.test requires fsverity to be configured in the kernel
and the fsverity-utils package to be installed.

- The evm_hmac.test requires specific EVM states.

For these reasons, limit running the kernel tests to the UML environment.

- Define a new configuration option named "--enable-kerneltests"
- Add ability of conditionally disabling the kernel tests from ci.  The
default is running all of the tests.

Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>