Add openssl-devel-engine package dependency for Fedora
Openssl 3.0 introduced the concept of "providers", deprecating "engine"
support. Consequently, in recent Fedora distros, the engine-related header
files have been moved away from the openssl-devel package, in
openssl-devel-engine package.

Fix this issue by including the latter as a dependency. Also, don't fail if
openssl-devel-engine does not exist yet, to accommodate the case of a
future transition of the CentOS distribution to splitting the packages.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>