Add SELinux package dependencies for evm_hmac.test on Fedora
Add policycoreutils and selinux-policy-targeted as package dependencies,
since they are required by evm_hmac.test.

Since those two packages were installed before, the likely reason of
lacking them for the test is that one of the packages listed in
ci/fedora.sh is not requiring them as a dependency anymore. Thus,
explicitly add them as a dependency for the test.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>