Move kernel-related tests to the test/kernel subdirectory
Separate user space tests from kernel tests, so that only the former group
of tests can be executed, if the intent is only to test ima-evm-utils.

Minimize the changes by placing the kernel tests in tests/kernel/, a
subdirectory of where the other tests are.

Update the relative path in the kernel tests, to reflect the fact that they
were moved one level down.

One side effect of this change is that gen-keys.sh is executed both in
tests/ and tests/kernel/, since it is required respectively by
sign_verify.test and fsverity.test. To make sure that keys are generated in
the directory where the script was invoked, remove the changing directory
command.

Finally, avoid to print the logs twice in build.sh.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>