Add get_private_key() and load_public_key() to functions_kernel.sh
Move get_private_key() and load_public_key() into the new library,
functions_kernel.sh.

Add the path of the temporary file that will hold a copy of the private key
as a new parameter of get_private_key(), and the path of the public key and
the target kernel keyring as new parameters of load_public_key().

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>