Move get_xattr() and check_load_ima_rule() to functions_kernel.sh
To avoid repeating the same code in individual tests, move the common
functions get_xattr() and check_load_ima_rule() into the new library,
function_kernel.sh.

Add the key path and mount point as additional parameters to
check_load_ima_rule(), so that the function can use the key for signing the
policy, and for creating the temporary file in the right place.

Finally, add function_kernel.sh to the shellcheck target.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>